@@ -1,16 +1,22 @@
-import './App.css';
-import FetchDataById from './component/FetchDataByIdUsingHooks/FetchDataById';
-import RegisterForm from './component/register-form/RegisterForm';
-
 function App() {
   return (
-    <div>
-      <RegisterForm />
+    <div className="h-screen bg-cyan-200 flex items-center justify-center">
+      <div className="w-80 bg-white p-5 rounded-xl text-center font-bold">
+        <h1 className="text-2xl font-bold">Simple Counter App</h1>
+        <h2 className="text-4xl my-3">{0}</h2>
+        <div className="">
+          <button className="btn bg-cyan-500 shadow-cyan-500/50">
+            Increment
+          </button>
+          <button className="btn bg-pink-500 shadow-pink-500/50">
+            Decrement
+          </button>
+          <button className="btn bg-indigo-500 shadow-indigo-500/50">
+            Reset
+          </button>
+        </div>
+      </div>
     </div>
-    
-    // <div className="App">
-    //   <FetchDataById />
-    // </div>
   );
 }
 
